Sram自检测成功 且速度非常快 但是需要去除GPIO72上的电阻

+++ Sources_Bsp/Bsp_AD5546.h
@@ -11,7 +11,7 @@
 extern "C" {
 #endif
 
-FP64 Bsp_AD5546HexToVolt(INT16U hex);
+FP32 Bsp_AD5546HexToVolt(INT16U hex);
 
 INT16U Bsp_AD5546VoltToHex(FP64 volt);
